requirements.txt
opencv-python>=4.5
numpy>=1.20
pupil-apriltags
pyserial
pyyaml
